Making AlternateTypeRule an interface for building out additional AlternateTypeRules
